Changed a lot
Worked on better console clarity, simplified things. deleted some old ones. Right now the user can select a steam folder and it checks if it is a steamfolder, it checks for if palserver is installed. it gets the users that were logged in at somepoint and stored in the loginusers.vdf. the user can select their steam acount in a dropdown and confirm it, then on serverside the right path for the palserver worlds gets created to read out the worlds there. world reading is not in play yet.