GUI Layout & Setup
worked a bit more on the general GUI Layout so that it represents more or less now the end product. The File Setup is now a little clearer. Added Auto Setup button for default installations (not fully implemented). World editor is just an html sketch for now.
Select Local Dir button is a dummy and could be in theory removed since palworld local savestates are allways located at %localappdata%